requirements.txt
dj-database-url==0.4.1
#Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==2.0.6
django-cms==3.4.1
djangocms-text-ckeditor
djangocms-file
djangocms-picture
djangocms-teaser
djangocms-link
django-image-cropping
